Split main.py into several files by functionality
